feat: running status table & frequency tip
